+++ products/views.py
@@ -403,8 +403,10 @@ def validate_edit_product(request, product_name):
         if product_form.is_valid():
             new_product = None
 
-            # Getting rid of the old image if a new one exists
-            if 'image' in request.FILES and os.path.exists(old_image_path):
+            # Getting rid of the old image if a new one exists in the workspace
+            if (settings.DEBUG
+                    and'image' in request.FILES
+                    and os.path.exists(old_image_path)):
                 os.remove(old_image_path)
 
             if different_form:
